Move exr_utils.py to core/utils folder
This was placed in the wrong folder due to a bad rebase.
