+++ src/ng_model_gym/core/utils/exr_utils.py
